app: lower query service impl
